+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * Added a `as.data.frame` method for `twoway` objects. This similifies graphical displays and other computations.
+* `print.twoway()` gets a `border=2` option to print the result in a fancy table with horizontal and vertical separators [thx: Richard Heiberger]